Fix: ADCs provide erroneous readings. This is due to the fact that the same pins can be used both as analog and digital inputs. The problem however is, that once pins are configured as digital input pins they no longer report correct values when used as analog inputs.
By their nature (as defined via the schematic) those pins are primarily
defined as analog inputs and are consequently used as such. A digital
input functionality is emulated on top of the analog one, by putting
the transfer point between LOW and HIGH at the halfway point of the
full analog voltage range.